fix: use int64 for block index comparison to prevent overflow

+++ weed/storage/erasure_coding/ec_locate.go
@@ -55,7 +55,7 @@ func LocateData(largeBlockLength, smallBlockLength int64, shardDatSize int64, of
 func moveToNextBlock(blockIndex int, isLargeBlock bool, nLargeBlockRows int64) (nextBlockIndex int, nextIsLargeBlock bool) {
 	nextBlockIndex = blockIndex + 1
 	nextIsLargeBlock = isLargeBlock
-	if isLargeBlock && nextBlockIndex == int(nLargeBlockRows)*DataShardsCount {
+	if isLargeBlock && int64(nextBlockIndex) == nLargeBlockRows*DataShardsCount {
 		nextIsLargeBlock = false
 		nextBlockIndex = 0
 	}
